+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Adding the `minClusterSize` option to clusterers.
 * Adding limited version of `metrics/distance/damerau-levenshtein`.
 * Adding limited version of `metrics/distance/levenshtein`.
+* Adding bitwise version of `metrics/distance/hamming`.
 * Adding normalized version of `metrics/distance/hamming`.
 * Moving `stats/ngrams` to `tokenizers/ngrams`.
 * Moving `keyers/omission` to `phonetics/omission`.